Trying to get top-down to meet bottom-up in middle
Basically a cross between trying to write something more interesting
than Hello World (though I really should have started there) and getting
it to actually run on the platform that I'm trying to build for running
the sort of App that I want to run on it.

Yes, this is a very confused piece of the puzzle at the moment.